requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
motor==3.1.1
pydantic==2.6.0
aiofiles==23.1.0
python-multipart==0.0.6
pytest==7.4.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
Pillow==10.2.0
